gh-699: bubble up graphql root without modifying it to Local.Get
There was a "filtersAndResolvers" helper type in use before, which was
most likely from the time when filters were extracted at a different
level than the class.

This is no longer needed, we can now simply bubble up the class. This
way we can access the requests loggers from the root map.