gh-788: rename GetMapped... to MustGetMapped... to indicate it panics
I'm not at all happy that these methods panic, however, they are widely
used throughout the app and refactoring them all to proper error
handling is not feasible at the moment.

Calling them MustGet... is at least an idiomatic way in Go to tell the
developer that this method panics if an error occurs. This allows us to
add a safe Get... for new code at least.

This doesn't really reduce the tech debt, but at least it makes it a bit
more obvious ... and prevents us from adding even more on top of it.