gh-801: remove old graphql introspection based tests
Why are we removing them?

- They have been turned off (with the env var) on CI for a while, so
right now they’re not adding any value

- The prototype is no longer the source of API changes. Newer changes
were implemented in weaviate first or in some cases only in weaviate

- One of the additional benefits of the tests was checkings that Object
names don’t collide in GraphQL. We have component tests in the graphql
packages that do that now

- With most of the resolvers being implemented we also have explicit
tests to see that the GQL endpoints work correctly end-to-end, i.e. both
structure and results