gh-647: move GetPropsOfType from schema manager to schema
Putting it into the manager was an incorrect assumption before. The
manager is just responsible for coordinating and storing changes.
However, all other read-only queries that access the schema are on the
schema-type itself. This way they can easily be consumed in the existing
connectors. Whereas the schema manager might not be accessible from
within the connectors.