gh-978 don't overwrite aggregation type as numerical
previously any numerical aggregation (including count) would always set
the property type to numerical. So a non-numerical aggregation, such as
a text-based one might sometimes end up as text (if count came first,
then text props later) and sometimes as numerical if the order was
reversed.

This also lead to flakyness in tests, however, the source for the flaky
tests were not the tests themselves, but actual flaky behavior