gh-1021 [skip ci] delete old importert tool, start fixing tests
previously the tests were relying on the importer to create a test
setup. However, the importer was relying a lot on the old patch
semantics, so instead of fixing it, we're fixign the test setup.

After this each test suite will do its own setup and tear down. This
commit starts with the test/acceptance/actions/ test suite. Others will
follow in future commits.

Skipping ci, as this is a WIP until all tests are fixed.