gh-1069 don't depud based on ID alone, dedup based on all 3
An ID makes an object unique in Weaviate, but it's still not enough to
only run deduplications based on it. This is because with
SelectProperties, i.e. the user-specified instructions to traverser the
graph, there is a difference. For example if A has a reference to B of
class "BestClass" and the user asks for all references from A of type
"WorstClass", we will send a request that won't have a result. If the
user now asks for all references of clasess "WorstClass", "BestClass"
these are clearly two different requests even with the same id (since we
don't know the class yet at resolve time.) If we were to dedup only on
ID (as we did before), we'd send only the first request (asking for
WorstClass) and would end up with no results.