[skip ci] gh-1148 add primitive vector cache for hnsw index
The hnsw index is most efficient if the vectors are held in memory.
However, we cannot assume that the user has infinite memory available.
Thus we can limit it with a simple cache. The cache is primitive in the
sense that it simply gets purged when it's full.

As part of the spikes preceeding this we have also investigated smarter
caches, that would purge entries with fewer recent hits first, etc.
However the orcehstration of that turned out to be more expensive than
the benefit it provided.

The cost of fillign the cache is not astronomically high, so this seems
to be a good solution for now.

The size is currenlty hard-coded to 10k items. This should get us far
enough on small datasets and still should have a decent performance even
if it gets purged a lot on medium sized set.

To support large sets (where it's a given that cluster nodes are not
tiny) ideally we'd make the limits dynamic. If there is still
allocatable RAM (keeping in mind that the key/value store is also
mmapped and needs mem), it could grow. This would especially make it
easier on datasets with a lot of indices/shards where we can't easily
predict which index/shard will take the highest load.