gh-1219 deal with deleted entrypoint in deletion cleanup
In regular searches we are currently just erroring. However, if the
cleanup works again due to this fix, it will be automatically fixed in
the next cycle.

A prettier solution would probably be to fix it right away in the same
cycle, but I don't know yet how common - if ever - this case is.