gh-1272 extract and refactor scan by doc id function
It was part of the inverted package, but I don't think it really belong
there. It was just one of the packages that were extracted early on, so
it was already "independent".

However, this issue introduced the DocID package, grouping all kinds of
docid related things. I think this is a great fit for the scanByDoc ID
function.

While moving it, I also split it up as it had become way too big.