gh-1358 don't use buffered channel in class batch writer
This could use to a situation where we would accept the cancel signal
while the buffer was still filled. This would in turn mean we terminate
the save operation with unsaved elements in the queue.

After discussion within the team, we don't see any reason for the
buffered channel (anymore), so we don't think we'll lose anything by
using a regular, unbuffered channel.

A new intergration test which uses a real `db.Repo` with the classifier
has been added as I could not reproduce the bug with a storage fake. I
still don't know what exactly it took to reproduce it, but I assume it
was a combination of timings that happen to be very present with a real
DB, but aren't present with the fake - although I added various delays
to the fake.

The new integration test had about a 50% failure rate prior to the fix,
but has now completed 100 subsequent runs succesfully. This makes me
reasonably confident that this indeed closes #1358.