gh-1459 dynamically load modules, including t2v-contextionary
This module was previously hard-coded. We also had to move the
validation around a little bit, since the config - which is the first
thing that is read when the app starts - now depends on the module
provider as well. So, we're running the final part of the validation
only after the module provider has started (but not yet initialized the
modules).