gh-1523 support disk segment cursor
There is no logic as of now to combine the results of multiple segments
and right now we assume there is either a disk segment or a memtable
present that we want to iterate over.

It will be subject of the following commits to merge the results and
strategically forward only the correct cursors when multiple are
present.