gh-1703 remove "combining.tmp" scratch files on startup
Such scratch files exist if a crash occured while two files were being
combined. The presence of such a file means that the process was not
finished and the file is not usable.

Prior to this fix, startup was blocked because the name of the file
couldn't be parsed. Thix fix does not alter the parsing, but actually
removes the file, knowing that it's not usable. The next combinig cycle
will simply create another one.

Note that this commit also changes the commit log switching threshold,
which helped discover this bug. Previously it would be switched at 500MB
and combined up to 1.75*500MB. While the latter number hasn't changed, a
switch now already occurs at 100MB, leading to smaller files and
slightly more frequent combining/condensing cycles. This improves
startup times after a crash, since the latest (uncondensed) log is now
smaller than it typically would be.

closes #1703