gh-1713 gh-1716 read LSM WAL buffered, pause writing WAL while reading
So we don't accidentally create more entries to be read next time while
we're reading. Since we're only reading from disk there is no need to
write anything to disk, as we don't need the "safety". If something
fails we can just read the same file again.

The above scenario led to a weird case of duplicating the logs over and
over again when an external factor (in this case a Kubernetes
initialDelaySeconds timeout) caused restarts. This then lead to more
data being parsed on the next run, making a success less likely with
each round.

In addition we are using a buffered reader when ingesting WALs, as there
was an incredibly high amount of syscalls because of the small chunks
being read.