gh-1732 cache int greater than filters
This required setting up a new row reader and in turn a new cursor
capability that retries only keys, but no values. This has been
implemented only for "Set" type so far and not yet for "Map", meaning we
do not yet support caching GreaterThan on frequency-types, e.g. strings