add additional tests to distributed transactions
while preparing a video for one of the upcoming features in sprint 3 I
made an assumption about our current transaction implementation leading
to tx open errors if nodes all think they were first due to network
delays. I wasn't sure if that scenario was actually covered and while we
had a similar looking test we did not have one that tested exactly this
scenario. This is why I added this to new test and to my delight it is
already passing.