gh-1868 properly propagate "clear links" across commit logs.
The following situation caused an issue. Imagine there are two commit
logs in chronological order:

Log 1:
- AddLink for Node/Level/Link 0/0/1
- AddLink for Node/Level/Link 0/0/2

Log 2:
- ClearLinks for Node/Level 0/0
- AddLink for Node/Level/Link 0/0/3
- AddLink for Node/Level/Link 0/0/4

It should be quite obvious that after reading all the logs back in the
correct order node 0 will have the connections [3,4] at level zero, as
there was a "clear" event in between.

However, prior to this fix we would have ended up with [1,2,3,4] because
the clear information got lost when condensing each log independently.
This fix makes sure that information that relates to changes from prior
commit logs is kept when condensing a commit log. The actual bug was
when reading (as part of condensing) as we used the clear only to clear
the links (which from the perspective of log 2 were already empty). This
led us to condensing log 2 to just "add 3, add 4", whereas the correctly
condensed log would have been "clear, add 3, add 4". This is the case
now.

To write the tests needed for this fix, it was required to add some
functional options to the commit logger to control the commit log size
thresholds. As of now these are not yet exposed to the user, but this
would also add value.

fixes #1868

Potential follow-up tasks:

1. Expose Commitlog configuration (e.g. thresholds) to user, so they can
   finetune them to their needs
2. Clean up the redundant and partly unused code, and the confusion
   between condensor/condensor2 and deserializer/deserializer2