Improve Dev-Server Script
This PR aims to improve the DX for the dev server script:

* global variables are no longer hard-coded, they are taken from the
  environment if not set. If not provided defaults are used.
* repetition is removed among the invidivual configurations by moving
  all variables that are the same for every config to the global level

This change is non-breaking and running the script without setting any
env vars has the same effects as before. However, you can now customize
it. For example to change the log format to JSON and to run with a
different data path you can simply run:

```
LOG_LEVEL=json DATA_PERSISTENCE_PATH=./some-other-path \
  tools/dev/run_dev_server.sh local-no-modules
```